worker: move HostErrorSet and ContractLock to their own packages
