+++ worker/worker.go
@@ -1130,32 +1130,3 @@ func (w *Worker) prepareUploadParams(ctx context.Context, bucket string, contrac
 	}
 	return up, nil
 }
-
-// A HostErrorSet is a collection of errors from various hosts.
-type HostErrorSet map[types.PublicKey]error
-
-// NumGouging returns numbers of host that errored out due to price gouging.
-func (hes HostErrorSet) NumGouging() (n int) {
-	for _, he := range hes {
-		if errors.Is(he, gouging.ErrPriceTableGouging) {
-			n++
-		}
-	}
-	return
-}
-
-// Error implements error.
-func (hes HostErrorSet) Error() string {
-	if len(hes) == 0 {
-		return ""
-	}
-
-	var strs []string
-	for hk, he := range hes {
-		strs = append(strs, fmt.Sprintf("%x: %v", hk[:4], he.Error()))
-	}
-
-	// include a leading newline so that the first error isn't printed on the
-	// same line as the error context
-	return "\n" + strings.Join(strs, "\n")
-}
